bff/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "bff"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
